configure sensor

+++ src/main/java/pw/stud/envparam/model/SensorConfigRequest.java
@@ -4,8 +4,8 @@ public class SensorConfigRequest {
     int sensorId;
     String sensorName;
     Integer measurementFreq;
-    Float temperatureMax;
-    Float temperatureMin;
+    Double temperatureMax;
+    Double temperatureMin;
     Integer humidityMax;
     Integer humidityMin;
 
@@ -33,19 +33,19 @@ public class SensorConfigRequest {
         this.measurementFreq = measurementFreq;
     }
 
-    public Float getTemperatureMax() {
+    public Double getTemperatureMax() {
         return temperatureMax;
     }
 
-    public void setTemperatureMax(Float temperatureMax) {
+    public void setTemperatureMax(Double temperatureMax) {
         this.temperatureMax = temperatureMax;
     }
 
-    public Float getTemperatureMin() {
+    public Double getTemperatureMin() {
         return temperatureMin;
     }
 
-    public void setTemperatureMin(Float temperatureMin) {
+    public void setTemperatureMin(Double temperatureMin) {
         this.temperatureMin = temperatureMin;
     }
